refactor: share the swap resume/cancel actions between both hosts
MemoryPanel and SwapResumeHost each carried a byte-equivalent error reporter and the same resume/cancel orchestration down to the already-completed toast, differing only in the pre-step each runs first. The hook owns the engine calls and the reporting contract; the dialog close and the navigate stay at their call sites.
